Fix possible string truncation
As GCC 8.1 points out, we're at one point trying to write to a buffer of
size "PATH_SZ" the contents of another buffer of size "PATH_SZ". That's
pretty obviously a possible string truncation, which can be avoided
easily.

While we're at it, use "sizeof" instead of hardcoding some number.

And finally, snprintf() is already NUL-terminated, we don't have to do
that by ourselves.